+++ source/css/base/create.css
@@ -3,7 +3,7 @@
     flex-flow: row wrap;
     justify-content: space-between;
     padding-top: 20px;
-    width: 80%;
+    width: 70%;
     margin: 0 auto;
 }
