+++ source/js/components/overview/QucikNote.js
